added compile switch for the "blink" and regular program, started debugging the data decompression logics, tim2 and dma setup, dma doesn't work so far, tim2 pwm frequency is correct!

+++ Inc/led_timer.h
@@ -157,6 +157,6 @@ void sendToLed();
 extern RGBStream * frame;
 extern uint32_t clr_cnt;
 extern uint32_t bit_cnt;
-extern uint32_t * rawdata_ptr;
+extern uint8_t * rawdata_ptr;
 
 #endif /* LED_TIMER_H_ */
